+++ src/main/java/svenhjol/charm/Charm.java
@@ -1,6 +1,8 @@
 package svenhjol.charm;
 
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import svenhjol.charm.base.CharmClient;
 import svenhjol.charm.base.CharmCompat;
 import svenhjol.charm.base.CharmMessages;
 import svenhjol.charm.base.CharmSounds;
@@ -13,6 +15,7 @@ public class Charm extends MesonInstance {
     public static final String MOD_ID = "charm";
     public static LogHandler LOG = new LogHandler(Charm.MOD_ID);
     public static IQuarkCompat quarkCompat;
+    public static CharmClient client;
 
     public Charm() {
         super(Charm.MOD_ID, LOG);
@@ -20,5 +23,11 @@ public class Charm extends MesonInstance {
         CharmMessages.init(this);
         CharmSounds.init(this);
         CharmCompat.init(this);
+    }
+
+    @Override
+    public void onClientSetup(FMLClientSetupEvent event) {
+        super.onClientSetup(event);
+        client = new CharmClient();
     }
 }