+++ src/main/java/svenhjol/charm/base/CharmMessages.java
@@ -1,5 +1,7 @@
 package svenhjol.charm.base;
 
+import svenhjol.charm.base.message.ClientUpdatePlayerState;
+import svenhjol.charm.base.message.ServerUpdatePlayerState;
 import svenhjol.charm.brewing.message.ClientCakeAction;
 import svenhjol.charm.tools.message.ClientGlowingAction;
 import svenhjol.meson.MesonInstance;
@@ -11,5 +13,8 @@ public class CharmMessages {
 
         handler.register(ClientCakeAction.class, ClientCakeAction::encode, ClientCakeAction::decode, ClientCakeAction.Handler::handle);
         handler.register(ClientGlowingAction.class, ClientGlowingAction::encode, ClientGlowingAction::decode, ClientGlowingAction.Handler::handle);
+        //handler.register(ClientRunePortalAction.class, ClientRunePortalAction::encode, ClientRunePortalAction::decode, ClientRunePortalAction.Handler::handle);
+        handler.register(ClientUpdatePlayerState.class, ClientUpdatePlayerState::encode, ClientUpdatePlayerState::decode, ClientUpdatePlayerState.Handler::handle);
+        handler.register(ServerUpdatePlayerState.class, ServerUpdatePlayerState::encode, ServerUpdatePlayerState::decode, ServerUpdatePlayerState.Handler::handle);
     }
 }